only use major for rhel epel link

+++ getdess.sh
@@ -82,7 +82,9 @@ pre_install () {
 install_dependencies () {
   $pkg_man -y update
   for pkg in $packages; do
-    $pkg_man -y install "$pkg"
+    if ! command_exists "$pkg"; then
+      $pkg_man -y install "$pkg"
+    fi
   done
 }
 
@@ -90,7 +92,7 @@ install_certbot () {
   case "$os_release" in
     centos) echo y | $pkg_man -y install epel-release;;
     amazon) echo y | amazon-linux-extras install epel;;
-    rhel) echo y | $pkg_man -y install "https://dl.fedoraproject.org/pub/epel/epel-release-latest-$os_id.noarch.rpm";;
+    rhel) echo y | $pkg_man -y install "https://dl.fedoraproject.org/pub/epel/epel-release-latest-$($os_id | awk -F. '{print $1}').noarch.rpm";;
     *);;
   esac
   # Versions using this approach
